RUE-1976: materialize literal constants through one owner beside the integer kernel
`materialize_int_const` and `materialize_float_const` in `sema/analyze_ops.rs`
are now the only places an integer or floating-point constant becomes an AIR
`Const`. Ordinary literals, negated literals, struct-initializer fields,
`@ptr_write` values, named constant uses, and comptime-block results all route
through them, so one range rule, one payload encoding, and one "does not fit"
report stand behind every position.

That report is E0800 everywhere. Spec 6.5:5 already gives that code to a
constant value outside its type's range and reserves E1200 (6.5:12) for an
initializer that would trap, and `analyze_literal` was the position already
reporting it; the comptime block, the constant initializer, and the semantic
nucleus reported E1200 or split the same fact by sign. `let x: i8 = -129`,
`comptime { -129 }`, and `const A: i8 = -129` now give one code and one
message, naming the value the literal denotes rather than the magnitude its
operand spells. Trapping arithmetic keeps E1200. 3.1:17 and 3.1:18 say so, and
`types/integers.toml` pins each position.

Kernel adapters: `Type::integer_semantics` is the one match producing an
`IntegerType` from a `Type`, and `durable_int_width` the one producing it from
a `DurableType`. The negation minimum table, the pattern `wrapping_neg` sites,
`Type::literal_fits`, `Type::negated_literal_fits`, `const_int_fits`, the
nucleus fits table, and the body provider's width enumerations are gone in
favor of them, and api_inventory guards in rue-air and rue-compiler hold both.

`ErrorKind::LiteralOutOfRange` carries its value as `i128` so a negative one
renders; the `--error-format json` schema is unchanged.

+++ crates/rue-compiler/src/durable_comptime/host.rs
@@ -1704,7 +1704,7 @@ impl<A: DurableComptimeHostAuthority + ?Sized> rue_air::ComptimeRejections
 
     fn literal_out_of_range(
         &self,
-        value: u64,
+        value: i128,
         ty: &Self::Type,
         site: &rue_air::ComptimeDiagnosticSite<Self::ProgramKey>,
     ) -> Self::Failure {
